Added debounced number editor
Usually the number editor will call its update function every single
time the value changes. If you drag fast enough, that can be dozens of
times per second. This behaviour isn't ideal for us, because we
recalculate all of the graph data every time the data changes, so this
component wraps the number editor, so that it updates a local state
instead of calling the update function.

The update function is then debounced to only occur every x milliseconds.